+++ qr-menu/src/components/Header.jsx
@@ -1,0 +1,8 @@
+function Header(){
+    return(
+        <>
+        <h1>QR-MENU</h1>
+        </>
+    )
+}
+export default Header;
