+++ qr-menu/src/layouts/Dashboard.jsx
@@ -1,9 +1,10 @@
-import { Link, Outlet } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
 import { Suspense } from 'react';
-import { useURL } from '../hooks/useURL';
 
+import Header from '../components/Header';
+import Navigation from '../components/Navigation';
 const Dashboard = () => {
-  const url = useURL();
+
   return (
     <div
       style={{
@@ -14,18 +15,8 @@ const Dashboard = () => {
       }}
     >
       <header>
-        <h1>QR-MENU</h1>
-        <nav
-          style={{
-            display: 'flex',
-            flexDirection: 'row',
-            gap: '30px',
-          }}
-        >
-          <Link to={url.Home.path}>Home</Link>
-          <Link to={url.ModMenu.path}> Mod menu</Link>
-          <Link to={url.Menu.path}> Menu</Link>
-        </nav>
+      <Header/>
+       <Navigation/>
       </header>
       <main style={{ flex: 1 }}>
         <Suspense>
