+++ qr-menu/src/pages/Menu.jsx
@@ -1,8 +1,12 @@
+
 const Menu = () => {
-  return (
-    <div style={{ backgroundColor: 'lightgray', padding: '30px' }}>
+  return (<>
+ 
+   <div style={{ backgroundColor: 'lightgray', padding: '30px' }}>
       <h2>Menu</h2>
     </div>
+  </>
+   
   );
 };
